コース別ページを app ディレクトリに移動
- app ディレクトリ配下にも tailwindcss が効くよう修正
